Move sharding_utils.py to py/vttest
